refactor: merge `extenders.ts` into `functions.ts`
