+++ src/commands/General/shorturl.ts
@@ -3,7 +3,7 @@ import {
 	getLocalizedDesc,
 	replyError,
 	translateContext,
-} from "@/helpers/extenders.js";
+} from "@/helpers/functions.js";
 import { CommandData, SlashCommandProps } from "@/types.js";
 import { createEmbed } from "@/utils/create-embed.js";
 import useClient from "@/utils/use-client.js";
