+++ src/services/languages/locales/en-US/misc.json
@@ -13,9 +13,9 @@
 	"USER_NOT_FOUND": "Member/user not found!",
 	"NO_ARGS": "No arguments required.",
 	"NSFW_COMMAND": "This command can only be used in an NSFW channel.",
-	"OPTION_NAN_ALL": "Please specify an integer greater than 0 or `all`.",
 	"OWNER_ONLY": "Only the bot owner can use this command.",
 	"SELECT_CANCELED": "Selection canceled.",
+	"SELECT_PLACEHOLDER": "Select something",
 	"STATS_FOOTER": "● [Support Server]({{supportLink}})\n● [Invite JaBa to Your Server]({{inviteLink}})",
 	"TIMED_OUT": "Time out.",
 	"JUMP_TO_PAGE": "Specify the page you want to jump to (maximum **{{length}}**):",
