+++ src/services/languages/locales/ru-RU/misc.json
@@ -12,7 +12,6 @@
 	"MORE_THAN_ZERO": "Укажите целое число больше 0",
 	"NO_ARGS": "Агрументы не требуются",
 	"NSFW_COMMAND": "Данную команду можно использовать только в NSFW канале",
-	"OPTION_NAN_ALL": "Укажите целое число больше 0 или `all`",
 	"OWNER_ONLY": "Данную команду может использовать только владелец бота",
 	"SELECT_CANCELED": "Выбор отменён",
 	"STATS_FOOTER": "● [Сервер поддержки]({{supportLink}})\n● [Пригласить JaBa на свой сервер]({{inviteLink}})",
