+++ src/services/languages/locales/uk-UA/misc.json
@@ -12,7 +12,6 @@
 	"MORE_THAN_ZERO": "Вкажіть ціле число більше 0",
 	"NO_ARGS": "Агрументи не потрібні",
 	"NSFW_COMMAND": "Цю команду можна використовувати тільки в NSFW каналі",
-	"OPTION_NAN_ALL": "Вкажіть ціле число більше 0 або `all`",
 	"OWNER_ONLY": "Цю команду може використовувати тільки власник бота",
 	"SELECT_CANCELED": "Вибір скасовано",
 	"STATS_FOOTER": "● [Сервер підтримки]({{supportLink}})\n● [Запросити JaBa на свій сервер]({{inviteLink}})",
